Query now supports returning maps
Since mid 2019 Datomic supports returning maps and since Datahike
wants to map Datomics Query-API I added it. The most changes were
made on the datalog-parser:
https://github.com/lambdaforge/datalog-parser/pull/9

The changes made to datahike are a function convert-to-return-maps
that takes a set and maps the given keys on its values and returns
a vector of the created maps. The changes made to the function q
are - besides formatting - a check if returnmaps are parsed and if
so calling aforemention function to convert the result.

Besides adding another parse-step and the check if there are
returnmaps parsed there should not be too much of a performance
impact as far as I can see it. The behaviour should now map
Datomics.

- Closes #149
- Depends on https://github.com/lambdaforge/datalog-parser/pull/9

Thanks to vlaad for bringing this up