Implemented automated releases to Clojars
Since our goal is to release more often I implemented a first draft
to automate the releases. Releases are a very individual thing and
I tried to cover datahikes previous manual process as far as
possible. This means that a commit to development creates a release
if there is a SNAPSHOT-version specified in project.clj and a
commit to master creates a release if there is a non-SNAPSHOT-version
specified in project.clj.

Besides that there is some streamlining of the steps happening in
the usual testing and building procedure. Furthermore I introduced
the new workflow-feature of CircleCI and now the steps are able to
run concurrently. This seems to be quite a bit faster.

The decision is to keep it simple but avoiding manual work in the
future. Therefor I am sticking with CircleCI even if there are
other good options. Another decision is to continue keeping the
current version in project.clj. This has downsides since you can
not easily tag a commit with git and get a new release because you
would have to adjust the version manually in project.clj and tag
it. This is error prone.

Either tagging continues to be manually or it is solved with a
git hook. That way is described in doc/release.md. It is not the
nicest way to do it but I did not want to use a leiningen plugin
or upload a write-key to CircleCI. There should be sufficient
documentation in the release.md.

- Refers https://github.com/replikativ/datahike/issues/167