Create general Datahike JDBC backend
This commit includes a helper function that reads a map
from environ. It works in the case of the integration
test of datahike-jdbc
https://github.com/replikativ/datahike-jdbc/blob/development/bin/run-integration-tests#L32

Furthermore I am introducing timbre for logging so that
we can now see via info level log message what configuration
we are loading. That is helpful for more complex use
cases in JDBC.

- Closes https://github.com/replikativ/datahike/issues/161
- Needs to be merged together with
  https://github.com/replikativ/datahike-jdbc/pull/1
- introduces timbre as dependency