Configure database name with environ
For a use case where there are multiple databases running like in
datahike-server there it is useful to have named databases to
specify the database you want to use in each operation. This
feature was added but there was no way to specify the names in
using environment or others, only via config file. I added the
option to use environ to specify names equal to the other config
values.

Besides that I set the logging of the config to debug-level
because it is logging each time the config is called for exists?,
 create and connect and each time maybe with a different name.
That is confusing when name is not set and I want it to not
confuse users. Documenation is adapted.

- Closes #217