Add token authentication
Token authentication in it's basic form works. You can set a token in
the configuration and use this token to authenticate against the API.
There is the option to use dev-mode which authenticates you always
without token. Authentication is good to be used if you want to give
people the opportunity to run an instance e.g. from the Azure Button.

Dev mode might be helpful for future use cases as well as to avoid
complicated setup when only using locally. To avoid widespread
problems in public space dev-mode is not set as default and needs to
be activated explicitly in the config.

- Closes #16
- adds token authentication
- token can be set via configuration
- dev-mode disables authentication
- automated basic integration tests
- added unittests