Fixes purging history after changes due to upsert.
The issue was caused by temporal indices also storing current datoms (due to
the design change caused by the need to improve performance of
'upsert').
To fix the issue, we first had to remove the current datoms from the
temporal ones and then apply the purge algo that was already used before
'upsert'.