Implement lazy sequence for transact
The actual implementation is not complex since one can just feed
a sequence to Datahike. It was just a matter of accepting this.
The connector namespace is now a bit slimmer since I abstained
from using a multimethod. Instead I am checking the input data
in a simple cond.

Adapted the api documentation to match Datomic so only a map with
:tx-data as a key is allowed. Passing a vector or sequence is
still working but not documented anymore. The examples from the
documentation are added to the test.api namespace for testing if
our API works.

- Refers #196
- Closes #151
- Closes #78